pipeline/requirements.txt

requests
ratelimit
neo4j-driver

